Add more dynamism to the test
* Inner class now implements Runnable: this allows to remove the bridge method, as the inner class does not return any value anymore
* Dump the locals into an arraylist that is given to the inner class

Issues:
* The synchronization is not good
* The increment has to be kept in the inner loop, while it is also deported to the inner class at the moment

+++ src/fr/yimgo/testasm/SequentialSqrt.java
@@ -2,11 +2,11 @@ package fr.yimgo.testasm;
 
 public class SequentialSqrt {
   public SequentialSqrt() { }
-  public Double sqrt(int n) {
+  public Double sqrt(Integer n) {
     Double sum = new Double(0);
 
-    for (int i = 0; i < n; i += 1) {
-      sum += Double.valueOf(Math.sqrt((double) i));
+    for (Integer i = 0; i < n; i += 1) {
+      sum += Math.sqrt(i);
     }
 
     return sum;
